diseño acabado

+++ backend/src/routes/tasks.routes.js
@@ -4,6 +4,8 @@ const Task = require("../models/Task");
 const auth = require("../middleware/auth.midleware");
 const multer = require("multer");
 const Tag = require("../models/Tag");
+const User = require("../models/user.model");
+const enviarCorreo = require("../services/email.services");
 
 const storage = multer.diskStorage({
   destination: (req, file, cb) => {
@@ -58,6 +60,7 @@ router.post("/tasks", auth, upload.single("image"), async (req, res) => {
       userId: req.userId,
       image: req.file ? req.file.filename : null,
       tags: tagIds,
+      userRol: req.userRol,
     });
 
     await newTask.save();
@@ -92,7 +95,6 @@ router.put("/tasks/:id/add-tag", auth, async (req, res) => {
     const task = await Task.findById(id);
     if (!task) return res.status(404).json({ error: "Tarea no encontrada" });
 
-    // evitar duplicados
     if (!task.tags.includes(tagId)) {
       task.tags.push(tagId);
     }
@@ -123,10 +125,40 @@ router.put("/tasks/:id/remove-tag", auth, async (req, res) => {
   }
 });
 
-router.delete("/tasks/:id", async (req, res) => {
+router.delete("/tasks/:id", auth, async (req, res) => {
   try {
     const { id } = req.params;
+    const { avisar } = req.body;
+
+    const task = await Task.findById(id);
+
+    if (!task) {
+      return res.status(404).json({ error: "Tarea no encontrada" });
+    }
+    if (
+      req.userRol !== "administrador" &&
+      task.userId.toString() !== req.userId
+    ) {
+      return res.status(403).json({ error: "No autorizado" });
+    }
+
+    const user = await User.findById(task.userId);
+
     await Task.findByIdAndDelete(id);
+
+    if (
+      req.userRol === "administrador" &&
+      avisar &&
+      task.userRol === "user" &&
+      user?.email
+    ) {
+      await enviarCorreo(
+        user.email,
+        "Tarea eliminada",
+        `Hola ${user.name}, tu tarea "${task.title}" ha sido eliminada por un administrador.`,
+      );
+    }
+
     res.json({ message: "Tarea eliminada" });
   } catch (error) {
     res.status(500).json({ error: "Error al eliminar la tarea" });
